resume_parser/requirements.txt
openai>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
requests>=2.31.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
pydantic>=2.0.0 